You're in the hole and want to volunteer for more work?
Shop Visit: When you open the Reward Shop, the app checks two things:

Are you in debt?

Can you afford the cheapest available reward?

"Volunteer for Task" Button: If you are not in debt, but you cannot afford even the cheapest item, a new section appears at the bottom of the shop. It offers you the chance to "Volunteer for Task."

The Catch: This isn't free. As the button states, you are volunteering for an immediate, random task, but it is only worth HALF POINTS.

Availability: The button is only clickable if you do not currently have an active task or forfeit.

Disappearance: As soon as you complete tasks and earn enough points to afford any item, this "volunteer" option will disappear from the shop.